+++ src/app/hook/use-zandi.ts
@@ -1,4 +1,3 @@
-// hooks/useZandi.ts
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
